+++ lib/preval/visitors/micro.rb
@@ -37,21 +37,33 @@ module Preval
       end
 
       def on_method_add_arg(node)
-        if node.type_match?(:call, :arg_paren) &&
-           node.body[0].type_match?(:vcall, :@period, :@ident) &&
-           node.dig(0, 2).body == 'gsub'
+        if node.type_match?(:call, :arg_paren)
+          if node.body[0].type_match?(:vcall, :@period, :@ident) &&
+             node.dig(0, 2).body == 'gsub'
 
-          left = node.dig(1, 0, 0, 0, 1)
-          right = node.dig(1, 0, 0, 1)
+            left = node.dig(1, 0, 0, 0, 1)
+            right = node.dig(1, 0, 0, 1)
 
-          if left.is?(:string_literal) &&
-             right.is?(:string_literal) &&
-             [left, right].all? do |node|
-               node.dig(0, 1).is?(:@tstring_content) &&
-               node.dig(0, 1).body.length == 1
-             end
+            if left.is?(:string_literal) &&
+               right.is?(:string_literal) &&
+               [left, right].all? do |node|
+                 node.dig(0, 1).is?(:@tstring_content) &&
+                 node.dig(0, 1).body.length == 1
+               end
 
-            node.dig(0, 2).update(:@ident, 'tr')
+              node.dig(0, 2).update(:@ident, 'tr')
+            end
+          elsif node.dig(0).type_match?(:method_add_block, :@period, :@ident) &&
+                node.dig(0, 0, 0).type_match?(%i[array vcall], :@period, :@ident) &&
+                node.dig(0, 0, 0, 2).body == 'map' &&
+                node.dig(0, 2).body == 'flatten' &&
+                node.dig(1).is?(:arg_paren) &&
+                node.dig(1, 0, 0, 0).is?(:args_new) &&
+                node.dig(1, 0, 0, 1).is?(:@int) &&
+                node.dig(1, 0, 0, 1).body == '1'
+
+            node.dig(0, 0, 0, 2).update(:@ident, 'flat_map')
+            node.update(:method_add_block, node.dig(0, 0).body)
           end
         end
       end