+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Fold constant for exponentiation if exponent is 0 and value is an integer.
 - Replace `.reverse.each` usage with `.reverse_each`.
 - Replace `foo ... in` loops with `.each do` loops.
+- Replace `.gsub('...', '...')` with `.tr('...', '...')` if the arguments are strings and they are of length 1.
 
 ## [0.2.0] - 2019-04-18
 ### Added