+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Replace `.gsub('...', '...')` with `.tr('...', '...')` if the arguments are strings and they are of length 1.
 - Replace `def foo; @foo; end` with `attr_reader :foo`.
 - Replace `.shuffle.first` with `.sample`.
+- Replace `.map { ... }.flatten(1)` with `.flat_map { ... }`.
 
 ## [0.2.0] - 2019-04-18
 ### Added